+++ src/main/java/com/example/practice/practice2/springIoc/IEncoder.java
@@ -1,0 +1,6 @@
+package com.example.practice.practice2.springIoc;
+
+public interface IEncoder {
+    String encodeString(String message);
+    byte[] encodeByte(String message);
+}
