+++ src/main/java/com/example/practice/practice2/springIoc2/Encoder2.java
@@ -1,0 +1,19 @@
+package com.example.practice.practice2.springIoc2;
+
+public class Encoder2 implements IEncoder2 {
+
+    private IEncoder2 iEncoder2;
+
+    public Encoder2(IEncoder2 iEncoder) {
+        this.iEncoder2 = iEncoder;
+    }
+    @Override
+    public String encodeStr(String message) {
+        return iEncoder2.encodeStr(message);
+    }
+
+    @Override
+    public byte[] encodeByte(String message) {
+        return iEncoder2.encodeByte(message);
+    }
+}
