+++ src/main/java/com/example/practice/practice2/validation/ValidationTestApplication.java
@@ -1,0 +1,11 @@
+package com.example.practice.practice2.validation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ValidationTestApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ValidationTestApplication.class, args);
+    }
+}
