+++ src/main/java/com/example/practice/practice2/validation/annotation/YearMonth.java
@@ -1,0 +1,30 @@
+package com.example.practice.practice2.validation.annotation;
+
+import com.example.practice.practice2.validation.validator.YearMonthValidator;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.*;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target({ METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER, TYPE_USE })
+@Retention(RUNTIME)
+@Constraint(validatedBy = { YearMonthValidator.class })
+public @interface YearMonth {
+    String message() default "형식에 맞지 않습니다.";
+
+    /**
+     * @return the groups the constraint belongs to
+     */
+    Class<?>[] groups() default { };
+
+    /**
+     * @return the payload associated to the constraint
+     */
+    Class<? extends Payload>[] payload() default { };
+
+    String pattern() default "yyyyMMdd";
+}
