+++ src/main/java/com/example/practice/practice2/validation/validator/YearMonthValidator.java
@@ -1,0 +1,34 @@
+package com.example.practice.practice2.validation.validator;
+
+
+import com.example.practice.practice2.validation.annotation.YearMonth;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class YearMonthValidator implements ConstraintValidator<YearMonth, String> {
+
+    private String pattern;
+
+    @Override
+    public void initialize(YearMonth constraintAnnotation) {
+        ConstraintValidator.super.initialize(constraintAnnotation);
+
+        // YearMonth 어노테이션에 있는 pattern메소드 값으로 세팅하여 사용
+        this.pattern = constraintAnnotation.pattern();
+    }
+
+    @Override
+    public boolean isValid(String value, ConstraintValidatorContext context) {
+        // pattern디폴트 값은 'YYYYMM'
+        try {
+            LocalDate.parse(value + "01", DateTimeFormatter.ofPattern(this.pattern));
+        } catch (Exception e) {
+            return false;
+        }
+        return true;
+    }
+
+}
