+++ src/main/java/com/example/practice/practice3/objectmapper/dto/Car.java
@@ -1,0 +1,31 @@
+package com.example.practice.practice3.objectmapper.dto;
+
+public class Car {
+    private String name;
+    private String carNumber;
+    private String type;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getCarNumber() {
+        return carNumber;
+    }
+
+    public void setCarNumber(String carNumber) {
+        this.carNumber = carNumber;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+}
