+++ src/main/java/com/example/practice/practice3/spring_exception/Exception처리.txt
@@ -1,0 +1,6 @@
+Spring에서 Exception처리하는 방법
+
+1. @ControllerAdvice(@RestControllerAdvice): Global 예외 처리 및 특정 package / controller 예외처리 가능
+2. @ExceptionHandler: 특정 controller 예외처리 가능
+
+
