+++ src/main/java/com/example/practice/practice3/spring_exception/SpringExceptionApplication.java
@@ -1,0 +1,10 @@
+package com.example.practice.practice3.spring_exception;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+@SpringBootApplication
+public class SpringExceptionApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(SpringExceptionApplication.class, args);
+    }
+}
