--- src/main/java/com/example/practice/springIoc/IEncoder.java
@@ -1,6 +1,0 @@
-package com.example.practice.springIoc;
-
-public interface IEncoder {
-    String encodeString(String message);
-    byte[] encodeByte(String message);
-}
