--- src/main/java/com/example/practice/springIoc2/IEncoder2.java
@@ -1,8 +1,0 @@
-package com.example.practice.springIoc2;
-
-import org.apache.logging.log4j.message.Message;
-
-public interface IEncoder2 {
-    String encodeStr(String message);
-    byte[] encodeByte(String message);
-}
