--- src/main/java/com/example/practice/validation/ValidationTestApplication.java
@@ -1,11 +1,0 @@
-package com.example.practice.validation;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ValidationTestApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ValidationTestApplication.class, args);
-    }
-}
