+++ src/main/resources/json/OMApplicationSample.json
@@ -1,0 +1,9 @@
+{
+  "name" : "홍길동",
+  "age" : 30,
+  "userCars" : [
+    "name" : "Avante",
+    "carNumber" : "123가1234",
+    "Type" : "세단"
+  ]
+}
